Solicitud de Adopcion fase 1 v2

+++ database/migrations/2014_10_12_000000_create_users_table.php
@@ -15,7 +15,7 @@ class CreateUsersTable extends Migration
     {
         Schema::create('users', function (Blueprint $table) {
             $table->increments('id');
-            $table->integer('ci')->unique();
+            $table->bigInteger('ci')->unique();
             $table->string('ci_extencion');
             $table->string('nombres');
             $table->string('apellido_paterno');
